Move pizza data, useArray,  calculate price functions to utils
