+++ src/utils/calculatePrice.spec.js
